Add benchmarks finel_clean vs clean after each pack
